pyproject.toml
[project]
name = "mingdao-mcp-login"
version = "0.1.0"
description = "明道云账号密码登录 → HAP 个人 MCP JSON（OAuth 集成）。"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "pycryptodome>=3.20.0",
  "requests>=2.31.0",
]

[project.scripts]
md-mingdao-login = "mingdao_mcp_login.md_login:entrypoint"
md-generate-mcp-config = "mingdao_mcp_login.generate_mcp_config:entrypoint"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mingdao_mcp_login"]
